+++ R/combo.R
@@ -16,6 +16,9 @@
 #' format. For example GSA \code{"8,9,10,11.1,11.2"} becomes \code{"891011"}.
 #'
 #' @seealso
+#' \code{\link{append.id}} modifies the \code{Assessment_ID} of an existing STAR
+#' object.
+#'
 #' \code{\link{lookup.species}} is a data frame containing species codes and
 #' names.
 #'